Slicing
Updated the sword attack to slice the projectiles in half.

+++ Assets/Scripts/CountdownTimer.cs
@@ -24,7 +24,7 @@ public class CountdownTimer : MonoBehaviour
         isCountDownComplete = true;
     }
 
-    public void StartTimer(int CountDownValue)
+    public void StartTimer(float CountDownValue)
     {
         isCountDownComplete = false;
         StartCoroutine(StartCountDown(CountDownValue));
